refactor(config): centralize migration orchestration with MigrationManager
Creates MigrationManager class to centralize all migration logic:
- migration-manager.ts: Single orchestration class with clear methods
  - needsMigration(): Check if migration is required
  - autoMigrate(): Silent migration for config loading
  - upgradeInteractive(): Interactive CLI upgrade
  - migrate(): Programmatic migration access

Updates:
- loader.ts: Uses MigrationManager.autoMigrate()
- init.ts: Uses MigrationManager.upgradeInteractive()
- Removes duplicated upgradeConfig function

Benefits:
- Single source of truth for migration logic
- Clearer separation between pure migration and orchestration
- More testable and maintainable
- Consistent migration behavior across the codebase
